requirements.txt
torch==2.5.1
torchvision==0.20.1
flask==2.3.2
flask-cors==3.0.10
onnxruntime==1.16.0
numpy==1.24.3
Pillow==9.5.0
onnx==1.17.0
opencv-python==4.11.0.86
matplotlib==3.7.1